requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-common-helpers==0.9.1
djangorestframework==3.6.3
funcsigs==1.0.2
idna==2.5
mock==2.0.0
oauthlib==2.0.2
pbr==3.1.1
pyenchant==1.6.8
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
twython==3.5.0
urllib3==1.21.1
